--- a/medical-image-viewer/package.json
+++ b/medical-image-viewer/package.json
@@ -19,6 +19,7 @@
   },
   "dependencies": {
     "@element-plus/icons-vue": "^2.3.1",
+    "@vueuse/core": "^13.5.0",
     "cornerstone-core": "^2.6.1",
     "dcmjs": "^0.43.1",
     "dicom-parser": "^1.8.21",