--- a/medical-image-viewer/src/styles/main.css
+++ b/medical-image-viewer/src/styles/main.css
@@ -105,4 +105,94 @@
   .print-full-width {
     @apply print:w-full;
   }
+
+  /* 性能优化工具类 */
+  .gpu-accelerated {
+    transform: translateZ(0);
+    will-change: transform;
+  }
+
+  .contain-layout {
+    contain: layout;
+  }
+
+  .contain-style {
+    contain: style;
+  }
+
+  .contain-paint {
+    contain: paint;
+  }
+
+  .contain-all {
+    contain: layout style paint;
+  }
+
+  /* 无障碍访问工具类 */
+  .sr-only {
+    position: absolute;
+    width: 1px;
+    height: 1px;
+    padding: 0;
+    margin: -1px;
+    overflow: hidden;
+    clip: rect(0, 0, 0, 0);
+    white-space: nowrap;
+    border: 0;
+  }
+
+  .focus-visible {
+    @apply outline-none ring-2 ring-primary-300 ring-offset-2;
+  }
+
+  /* 响应式文本工具类 */
+  .text-responsive {
+    @apply text-sm md:text-base;
+  }
+
+  .text-responsive-lg {
+    @apply text-base md:text-lg;
+  }
+
+  .text-responsive-xl {
+    @apply text-lg md:text-xl;
+  }
+
+  /* 间距响应式工具类 */
+  .p-responsive {
+    @apply p-2 md:p-4;
+  }
+
+  .m-responsive {
+    @apply m-2 md:m-4;
+  }
+
+  /* 医学影像专用工具类 */
+  .medical-text-primary {
+    @apply text-gray-900;
+  }
+
+  .medical-text-secondary {
+    @apply text-gray-600;
+  }
+
+  .medical-text-muted {
+    @apply text-gray-500;
+  }
+
+  .medical-border {
+    @apply border-gray-200;
+  }
+
+  .medical-hover {
+    @apply hover:bg-gray-50;
+  }
+
+  .medical-bg-secondary {
+    @apply bg-gray-50;
+  }
+
+  .medical-sidebar {
+    @apply bg-white border-gray-200;
+  }
 }