… medical-image-viewer/UI_OPTIMIZATION_REPORT.md
… medical-image-viewer/src/components/file/FileListComponent.vue
… medical-image-viewer/src/components/layout/EnhancedLayoutComponent.vue
… medical-image-viewer/src/components/tools/ToolbarComponent.vue
… medical-image-viewer/src/composables/useAccessibility.ts
… medical-image-viewer/src/composables/usePerformance.ts
… medical-image-viewer/src/composables/useResponsive.ts
… medical-image-viewer/src/config/ui-optimization.ts
… medical-image-viewer/src/tests/FileListComponent.test.ts
… medical-image-viewer/src/utils/ui-optimization.ts